frontend/requirements.txt
streamlit>=1.35.0
Pillow>=10.0.0
pymupdf>=1.24.0
openai-whisper>=20231117
paddleocr>=2.7.3
paddlepaddle>=2.6.1
ollama>=0.2.0
numpy>=1.26.0
